Time: 0 ms (100%), Space: 9.4 MB (16.57%) - LeetHub

+++ 0410-split-array-largest-sum/0410-split-array-largest-sum.cpp
@@ -1,18 +1,18 @@
 class Solution {
 public:
     int splitArray(vector<int>& nums, int m) {
-        long long l = *max_element(nums.begin(),nums.end());
-        long long r = accumulate(nums.begin(), nums.end(), 0LL);
+        int l = *max_element(nums.begin(),nums.end());
+        int r = accumulate(nums.begin(), nums.end(), 0LL);
         while(l < r) {
-            long long mid = l + (r - l) / 2;
+            int mid = l + (r - l) / 2;
             if(duable(mid, m, nums)) r = mid;
             else l = mid + 1;
         }
         return l;
     }
 
-    bool duable(long long mx, int m, vector<int>& nums) {
-        long long sum = 0;
+    bool duable(int mx, int m, vector<int>& nums) {
+        int sum = 0;
         for(int num : nums) {
             sum += num;
             if(sum > mx) {
